agents/requirements.txt
httpx==0.27.0
pydantic==2.8.2
pyyaml==6.0.1
typer[all]==0.12.3
rich==13.7.1
